requirements.txt
PyPDF2
openai
faiss-cpu
tqdm